Add `seshat:new/4` to attach extra info to a counter
Currently the extra info is only returned by `seshat:format/1`: it's
useful to attach information you might use for Prometheus labels.
Previously we used the counter's name as passed to `seshat:new/3` in
the place of this label.

Attaching extra info is useful for QQs in RabbitMQ for example. They
need to attach the queue name and vhost so that info can be exposed to
Prometheus.