Optimise seshat:counters/3
If someone uses this function, the list of counters to return
is likely very short, quite possibly it has 1 element.
Rather than building a map of all counters and then dropping
the irrelevant ones, only add the relevant counters in the first place.

Testing with RabbitMQ an rabbit_quorum_queue:open_files function,
which only asks for 1 metric, the avery execution went down form 7ms
to 1ms with this change.